// webapp/src/api/model/upgrade.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { UpgradeEffectsItem } from './upgradeEffectsItem';
import type { EUpgradeRestriction } from './eUpgradeRestriction';
import type { TagRequirement } from './tagRequirement';

export interface Upgrade {
  effects: UpgradeEffectsItem[];
  id?: string;
  name: string;
  restriction: EUpgradeRestriction;
  slots: number;
  tagRequirement: TagRequirement;
  vendorPrice: number;
}